README
Build your real toy car with an engine and wheels and fix the raspberry pi inside the car. Connect raspberry pi to a screen using the wire, clone the project, make sure that each gpio pin connects correctly as in the node.js file, open the HTML file, and control the car.